Fix 6DoF pose conventions, uncertainty calibration, and relocalization supervision
- Fix uncertainty NLL: divide by sigma^2 (not sigma), use log(sigma^2) per Gaussian NLL
- Fix gradient accumulation: use valid_step_count instead of batch_idx for zero_grad/step
  boundaries, add end-of-epoch step for remaining grads, NaN skip preserves accumulated grads
- Fix nuscenes_occworld_dataset: add .float() to occupancy tensors (were uint8)
- Fix midair/tartanair datasets: add missing 'domain_tag': 'sim' to __getitem__ return
- Add 10 regression tests covering all fixes

+++ dataset/nuscenes_occworld_dataset.py
@@ -238,8 +238,8 @@ class NuScenesOccWorldDataset(Dataset):
         future_poses = all_poses[n_hist:]
 
         return {
-            'history_occupancy': torch.from_numpy(np.stack(history_occ)),
-            'future_occupancy': torch.from_numpy(np.stack(future_occ)),
+            'history_occupancy': torch.from_numpy(np.stack(history_occ)).float(),
+            'future_occupancy': torch.from_numpy(np.stack(future_occ)).float(),
             'history_poses': torch.from_numpy(np.stack(history_poses)).float(),
             'future_poses': torch.from_numpy(np.stack(future_poses)).float(),
             'scene_name': sample_info['scene_name'],
